% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbunDq3.r
\name{D2_yhc}
\alias{D2_yhc}
\title{R code for q=2, 3-habitat mixture of rarefaction/extrapolation curves}
\usage{
D2_yhc(data, mm)
}
\arguments{
\item{mm}{an int matrix for number of individuals for x1,x2,x3}

\item{x1}{a vector/matrix/list of species abundance frequency}

\item{q}{a numerical value for Dq}
}
\value{
a vector/matrix/list with species relative abundance or detection probability distribution
}
\description{
R code for q=2, 3-habitat mixture of rarefaction/extrapolation curves
}
\examples{
data(Abudata)
data1<-Abudata$data
mm1<-Abudata$mm
Dq_in_yhc(data1[,1],data1[,2],data1[,3],mm1,0)
}
