% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbunDq3.r
\name{Dq_in_yhc}
\alias{Dq_in_yhc}
\title{R code for q=0,1  3-habitat mixture of rarefaction curves}
\usage{
Dq_in_yhc(x1, x2, x3, mm, q)
}
\arguments{
\item{x1}{a vector/matrix/list of species abundance frequency}

\item{mm}{an int matrix for number of individuals for x1,x2,x3}

\item{q}{a numerical value for Dq}
}
\value{
a vector/matrix/list with species relative abundance or detection probability distribution
}
\description{
R code for q=0,1  3-habitat mixture of rarefaction curves
}
\examples{
data(Abudata)
data1<-Abudata$data
mm1<-Abudata$mm
Dq_in_yhc(data1[,1],data1[,2],data1[,3],mm1,0)
}
