% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbunDq3.r
\name{ggmiNEXT3}
\alias{ggmiNEXT3}
\title{Plot function for output of miNEXT3.
\code{ggmiNEXT3} plot the output of miNEXT3 based on ggplot.}
\usage{
ggmiNEXT3(x)
}
\arguments{
\item{x}{output of miNEXT3.}
}
\value{
a ggplot object
}
\description{
Plot function for output of miNEXT3.
\code{ggmiNEXT3} plot the output of miNEXT3 based on ggplot.
}
\examples{
data(Abudata)
data1 <- Abudata$data
result <- miNEXT3(data1)
ggmiNEXT3(result)
}
