\name{miNEXT3-package}
\alias{miNEXT3-package}
\alias{miNEXT3}
\docType{package}
\title{\packageTitle{miNEXT3}}
\description{\packageDescription{miNEXT3}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{miNEXT3}
  \packageIndices{miNEXT3}

  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
  \packageAuthor{miNEXT3}
  
  Maintainer: \packageMaintainer{miNEXT3}
}
\references{
  This optional section can contain literature or other references for
  background information.
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
\seealso{
  Optional links to other man pages
}
\examples{
  ## Optional simple examples of the most important functions
  ## Use \dontrun{} around code to be shown but not executed
}
