% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miNEXT3_Abundance.r
\name{miNEXT3_Abu}
\alias{miNEXT3_Abu}
\title{R code for 3-habitat mixture of 3-habitat mixture of rarefaction/extrapolation curves}
\usage{
miNEXT3_Abu(data = NULL, knots = 10, size = NULL, nboots = 0)
}
\arguments{
\item{data}{a vector/matrix/list of species abundance frequency}

\item{boot}{}

\item{q}{a numerical value for Dq}
}
\value{
a vector/matrix/list with species relative abundance or detection probability distribution
}
\description{
R code for 3-habitat mixture of 3-habitat mixture of rarefaction/extrapolation curves
}
\examples{
data(Abudata)
data1<-Abudata$data
mm1<-Abudata$mm
Dq_in(data1[,1],data1[,2],data1[,3],mm1,0)
}
